requirements.txt
Pillow==9.2.0
inky==1.3.2
